Modified MainActivity to simply set the packageContext as "this".
Modified the Questions and Answers classes so that all of the findViewById methods were called in the onCreate method and would therefore only ever be called once per View during the lifecycle of the activity. This also involved the inclusion of several new global objects in each class which are then called elsewhere in the code instead of using findViewById again.

These changes were recommended by a mentor on the Google Developer Challenge course on udacity.com.